rename module from Relationizer::BigQuery::Standard to Relationizer::BigQuery
